requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-core>=0.1.0
google-generativeai>=0.8.0
duckdb>=1.0.0
pandas>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
